requirements.txt
Django==3.1.5
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-haystack==3.0
djangorestframework==3.12.2
Whoosh==2.7.4